requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
exceptiongroup==1.3.0
idna==3.10
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.7.1
multidict==6.5.1
packaging==25.0
pamqp==3.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyfiglet==1.0.3
Pygments==2.19.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.1.0
requests==2.32.4
rich==14.0.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
